+++ hamming/hamming.h
@@ -5,7 +5,7 @@
 
 namespace hamming
 {
-  int compute(const std::string& strand1, const std::string& strand2);
+  size_t compute(const std::string& strand1, const std::string& strand2);
 }
 
 #endif
